Refactor/#4: 마이페이지 더미데이터 반영
