+++ ie_um/src/pages/MyPage/MyLikes.jsx
@@ -1,22 +1,16 @@
 import { RiHeart3Fill } from 'react-icons/ri';
-import * as S from './Style/MyLikesStyle';
+import * as S from '../../styles/PostStyle';
 import { useState } from 'react';
 import MyPagination from '../../components/Pagination';
+import { DummyCommunity } from '../../constants/DummyData';
 
-const MyLikes = () => {
+const MyLikes = ({ posts = DummyCommunity }) => {
    const [activePage, setActivePage] = useState(1);
-
-   const MOCK = Array.from({ length: 20 }).map((_, i) => ({
-      id: i + 1,
-      title: '성공회대 근처 맛집 후기',
-      date: '2025.08.03',
-      like: 123,
-   }));
 
    const itemsPerPage = 7;
    const LastItem = activePage * itemsPerPage;
    const FirstItem = LastItem - itemsPerPage;
-   const currentItems = MOCK.slice(FirstItem, LastItem);
+   const currentItems = posts.slice(FirstItem, LastItem);
 
    const handlePageChange = (pageNumber) => {
       setActivePage(pageNumber);
@@ -28,7 +22,7 @@ const MyLikes = () => {
          <S.List>
             {currentItems.map((item) => (
                <li key={item.id}>
-                  <S.Row>
+                  <S.PostLink to={`/community/${item.id}`}>
                      <div>
                         <S.PostTitle>{item.title}</S.PostTitle>
                         <S.PostDate>{item.date}</S.PostDate>
@@ -36,9 +30,9 @@ const MyLikes = () => {
 
                      <S.HeartWrap>
                         <RiHeart3Fill size={11} />
-                        <S.HeartCount>{item.like}</S.HeartCount>
+                        <S.HeartCount>{item.heart}</S.HeartCount>
                      </S.HeartWrap>
-                  </S.Row>
+                  </S.PostLink>
                   <S.Divider />
                </li>
             ))}
@@ -47,7 +41,7 @@ const MyLikes = () => {
          <MyPagination
             activePage={activePage}
             itemsCountPerPage={itemsPerPage}
-            totalItemsCount={MOCK.length}
+            totalItemsCount={posts.length}
             pageRangeDisplayed={itemsPerPage}
             onChange={handlePageChange}
          />
